Add input validation to __init__ method of Lognormal class
This commit adds input validation to the __init__ method of the Lognormal class
to ensure that the `low_bound` and `up_bound` inputs are floats and that `up_bound`
is greater than `low_bound`. If either of these conditions is not met, a TypeError
or ValueError is raised with a descriptive error message. These checks help prevent
bugs and make the code more robust.